Added a step to add an initial commit
