+++ src/main.js
@@ -3,10 +3,10 @@ import fs from "fs";
 import ncp from "ncp";
 import path from "path";
 import { promisify } from "util";
-import execa from "execa";
 import Listr from "listr";
 import { projectInstall } from "pkg-install";
 import mkdirp from "mkdirp";
+import { gitAddandCommit, initGit } from "./utils";
 
 const access = promisify(fs.access);
 const copy = promisify(ncp);
@@ -15,16 +15,6 @@ async function copyTemplateFiles(options) {
   return copy(options.templateDirectory, options.targetDirectory, {
     clobber: false,
   });
-}
-
-async function initGit(options) {
-  const result = await execa("git", ["init"], {
-    cwd: options.targetDirectory,
-  });
-  if (result.failed) {
-    return Promise.reject(new Error("Failed to initialize git"));
-  }
-  return;
 }
 
 async function createDirectory(options) {
@@ -77,6 +67,11 @@ export async function createProject(options) {
         !options.runInstall
           ? "Pass --install to automatically install dependencies"
           : undefined,
+    },
+    {
+      title: "Making first commit",
+      task: () => gitAddandCommit(options),
+      enabled: () => options.git,
     },
   ]);
